-Use the correct MixinMinecraftServer
